For code completion we need explicit typing
The :ref binding does not report a meaningful type so we have to force the right type.
Examples in the docs are for the unbound/string variety and therefore assum a Ref implementation and automatically use .value.
To resolve a confusion with the global window object in the browser rename window and windows to winbox and winboxes.